Move Visit to owner package (it's aggregate root)
