+++ src/main/java/org/springframework/samples/petclinic/owner/Pet.java
@@ -32,7 +32,6 @@ import javax.persistence.Table;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.samples.petclinic.model.NamedEntity;
-import org.springframework.samples.petclinic.visit.Visit;
 
 /**
  * Simple business object representing a pet.
