+++ src/main/java/org/springframework/samples/petclinic/owner/Visit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.samples.petclinic.visit;
+package org.springframework.samples.petclinic.owner;
 
 import java.time.LocalDate;
