+++ src/test/java/org/springframework/samples/petclinic/owner/OwnerControllerTests.java
@@ -46,7 +46,6 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.samples.petclinic.visit.Visit;
 import org.springframework.test.web.servlet.MockMvc;
 
 /**
