+++ src/test/java/org/springframework/samples/petclinic/service/ClinicServiceTests.java
@@ -33,9 +33,9 @@ import org.springframework.samples.petclinic.owner.Owner;
 import org.springframework.samples.petclinic.owner.OwnerRepository;
 import org.springframework.samples.petclinic.owner.Pet;
 import org.springframework.samples.petclinic.owner.PetType;
+import org.springframework.samples.petclinic.owner.Visit;
 import org.springframework.samples.petclinic.vet.Vet;
 import org.springframework.samples.petclinic.vet.VetRepository;
-import org.springframework.samples.petclinic.visit.Visit;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
